Update SLA boundaries to use double
See https://github.com/micrometer-metrics/micrometer/pull/1909

Closes gh-20837

+++ spring-boot-project/spring-boot-actuator-autoconfigure/src/main/java/org/springframework/boot/actuate/autoconfigure/metrics/ServiceLevelAgreementBoundary.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2019 the original author or authors.
+ * Copyright 2012-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,8 +22,8 @@ import io.micrometer.core.instrument.Meter;
 
 /**
  * A service level agreement boundary for use when configuring Micrometer. Can be
- * specified as either a {@link Long} (applicable to timers and distribution summaries) or
- * a {@link Duration} (applicable to only timers).
+ * specified as either a {@link Double} (applicable to timers and distribution summaries)
+ * or a {@link Duration} (applicable to only timers).
  *
  * @author Phillip Webb
  * @since 2.0.0
@@ -42,17 +42,17 @@ public final class ServiceLevelAgreementBoundary {
 	 * @param meterType the meter type
 	 * @return the value or {@code null} if the value cannot be applied
 	 */
-	public Long getValue(Meter.Type meterType) {
+	public Double getValue(Meter.Type meterType) {
 		return this.value.getValue(meterType);
 	}
 
 	/**
-	 * Return a new {@link ServiceLevelAgreementBoundary} instance for the given long
+	 * Return a new {@link ServiceLevelAgreementBoundary} instance for the given double
 	 * value.
 	 * @param value the source value
 	 * @return a {@link ServiceLevelAgreementBoundary} instance
 	 */
-	public static ServiceLevelAgreementBoundary valueOf(long value) {
+	public static ServiceLevelAgreementBoundary valueOf(double value) {
 		return new ServiceLevelAgreementBoundary(MeterValue.valueOf(value));
 	}
